wip: delete trash file
